# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Boo Coding Challenge",
    "description": "REST API's for comments, votes and likes on a comment.Ideal flow for the implementation is \n<ls><li>Create two users</li><li>Create one comment targeting one user on behalf of another user.</li><li>Vote the comment using commentId.</li><li>Like or dislike the comment using separate APIs.</li><li>Like and dislike maintains no record about the source. </li><li>Increment and decrement action is included in the same API, for increment use value as 1, to decrement use value as -1.</li></ls>"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/profile/{profileId}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "description": "Endpoint to sign in a specific user",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/profile/createProfile/": {
      "post": {
        "tags": [
          "Profile"
        ],
        "description": "Endpoint to sign in a specific user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/user/createUser/": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to create one user.",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/user/getUser/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to retrieve one user.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Profile": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "A Martinez"
        },
        "description": {
          "type": "string",
          "example": "Adolph Larrue Martinez III."
        },
        "mbti": {
          "type": "string",
          "example": "ISFJ"
        },
        "enneagram": {
          "type": "string",
          "example": "9w3"
        },
        "variant": {
          "type": "string",
          "example": "sp/so"
        },
        "tritype": {
          "type": "number",
          "example": 725
        },
        "socionics": {
          "type": "string",
          "example": "SEE"
        },
        "sloan": {
          "type": "string",
          "example": "RCOEN"
        },
        "psyche": {
          "type": "string",
          "example": "FEVL"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        }
      }
    },
...
  }
}